requirements.txt
openai==1.34.0
pydantic==1.10.12
python-dotenv==1.0.1
slack_bolt==1.19.0